requirements.txt
openai-whisper>=20231117
numpy>=1.22.0
torch>=2.0.0
pydub>=0.25.1
ffmpeg-python>=0.2.0
librosa>=0.10.1
scikit-learn>=1.3.0
tqdm>=4.65.0 